AMD helper refactoring
